Move unittests
